api_service/requirements.txt
aioredis==2.0.1
elasticsearch[async]==7.14
backoff==2.0.1
fastapi==0.78.0
orjson==3.7.2
pydantic==1.9.1
uvicorn==0.17.6
uvloop==0.16.0
elasticsearch_dsl==7.4.0
fastapi-cache2==0.1.8
python-dotenv==0.20.0
pytest==6.1.2
pytest-asyncio==0.12.0
aiohttp==3.7.2
gunicorn==20.1.0
redis<=4.3.3
